[Ref]: 영화 자동 업데이트 관련 스케줄러 로직 분리
